Always log the charmcraft version when indicated.
Moved out the the version logging from Dispatcher's init to the
MessageHandler, which will always send it to the file (first thing after
configuring that part of the logging) and to the terminal when the mode
goes verbose.

This is part of #48, but not of all it. It's what was important,
thought, other two messages remain unshown because of initialization
order, but those two messages will go away (or at least change heavily)
when the bootstrap refactor is done (see #139).