Produce the default help text.
For this I had to add a 'common' attribute to commands, to flag those are of "common use" (the "common" term was UX used in the docs).

Not integrated yet, this is part of the [big adjust help process](https://github.com/canonical/charmcraft/pull/153).

This PR includes [this previous one](https://github.com/canonical/charmcraft/pull/154).