snap: strict -> classic confinement for Charmcraft snap
Classic snaps on core20 require python3.  Add this collection
of stage packages as its own part to separate stage filtering
used in the charmcraft part.

There is an apparent bug in Snacpraft where it will set the
bin/python3 to an invalid symlink in this scenario.  Hardcode
the correct symlink until we have a fix in place for Snapcraft.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>