charmcraft config: introduce support for bases keyword
- Support short-form syntax where a base without 'build-on'
  or 'run-on' implies build_on=base coupled with run_on=base.

- Support long-form syntax where explicit build-on and run-on
  lists are used to support any configuration combination.

- Support combined lists, where a short-form base item will
  be expanded into the long-form equivalent.

Note that this does not include unit testing for errors.  There
is a fair bit of other work to be done to improve error handling.

Note that this does not automatically set any implied architecture
in the Config, though an empty list of architectures would imply
the default (host's) is to be used.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>