charmcraft: don't leak PYTHONPATH if running as snap
Now that the snap is classic and the environment specifies PYTHONPATH,
ensure that it is scrubbed from the process in main such that it does
not leak into children processes.

Add a simple smoke test to Github Actions to sanity check snap.  This
will be replaced with more extensive integration tests in the future.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>