main: verify runtime environment is supported
Add env module with helpers for querying/validating environment:
- is_charmcraft_running_from_snap()
- is_charmcraft_running_in_developer_mode()
- is_charmcraft_running_in_supported_environment()
- ensure_environment_is_supported()

Call ensure_environment_is_supported() from main to verify that
the runtime environment is supported.  If not supported,
CHARMCRAFT_DEVELOPER=1 would be required to be set in environment
for development purposes.

Update main tests to bypass the "unsupported" configuration error
with an autouse fixture.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>