build & pack: add and use default ubuntu 20.04 bases configuration (CRAFT-266)
- Add a default base of Ubuntu 20.04 to match the behavior of the
strict snap which always reports as Ubuntu 20.04.  Upcoming work
will enable the use of `bases:` configuration in the charmcraft.yaml
to override this.

- Update pack command for bundles to use default bases configuration
as well.

- Remove legacy charm name format as it will no longer be used.

- Remove legacy manifest bases support as it will no longer be used.
Update associated unit tests.

Note that there is no deprecation warning for the use of default
bases, this will be added in a follow up PR.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>