providers: gracefully handle when LXD not installed (CRAFT-329)
When building, ensure LXD is installed if going to use a managed
environment.

When cleaning, check if LXD is installed before attempting to
clean project containers.

Add two interfaces in providers to support this:

- ensure_provider_is_available() to raise error if not available
(for cases like build where the provider may be required).

- is_provider_available() to check and return bool (for cases like
cleaning where the provider does not need to be available).

Future work can extend ensure_provider_is_available() to assist
the user in installation of the provider.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>