pack: fetch logs from managed instance on error (CRAFT-333)
Catch error when failing to pack a Charm, retrieving the logs from
the managed instance for easier inspection by the user.

- Update logger to write to /tmp/charmcraft.log so that it can be
readily retrieved by the host Charmcraft instance.  Provide this
path with a new env.get_managed_environment_log_path() interface.
Additionally, ensure the log file is truncated with mode='w' so
the instance logs are only for the last run.

- prefix logs in managed instance with "mm "

- add providers.capture_logs_from_instance() to providers to retrieve
  logs and send them to logger.debug()

- update build to use this mechanism

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>