providers: account for project path in instance name (CRAFT-334)
If Charmcraft were to build the same project [name] simultaneously in
two different directories, collisions on the containers in use may
occur.

To account for this, this PR uses the project directory's inode in the
container name.  This will ensure uniqueness for projects in multiple
paths, but also seamlessly handle project directory moves that don't
cross the filesystem boundary, the common case.

Note that 'clean' command will not affect other project directories,
so the clean must occur in the original directory.  To handle any
potentially unassociated containers, future work will include more
sweeping purge capabilities for the clean command.  This is not unique
to the introduction of the use of the directory inode, but is
exacerbated by it.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>