commands/build: output charms to cwd when using provider (CRAFT-359)
If building in project directory, use the project path as the working
directory. The output charms will be placed in the correct directory
without needing retrieval. If outputing to a directory other than the
charm project directory, we need to output the charm outside the
project directory and can retrieve it when complete. Use the home
directory for this output inside the container.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>