build: create charm builder tool
Extract the build routine from the Builder class and create a stand-alone
charm builder tool. Invoke the tool from the Builder class passing all the
appropriate arguments to build charms. This tool will be invoked from the
charm plugin after parts integration.

The payload venv is now created from the user site of a staging venv to
cache the previous installation and avoid reinstalling packages that are
already installed. This is done to avoid spending a long time building
packages when binary wheels are disabled.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>