providers: add support to install Charmcraft snap from store (CRAFT-502)
On Linux, the default behavior is to inject the host snap into the target
environment.

On other platforms, the Charmcraft snap is installed from the Snap Store.

When installing the snap from the Store, we check if the user specifies a
channel, using CHARMCRAFT_INSTALL_SNAP_CHANNEL=<channel>.  If unspecified,
we use the default determined by get_managed_environment_snap_channel_default().

On Linux, the user may specify this environment variable to force Charmcraft
to install the snap from the Store rather than inject the host snap.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>