tests: skip broken tests when running on Windows (CRAFT-504)
As a starting point for Windows testing, we disable all tests which fail
on Windows.  Over time we should update the tests to either work on
Windows, or mark them as specifically unsupported on Windows.

Tests are disabled entirely for the test_parts module, as craft-parts
does not yet support Windows.  In practice in may not be much effort
to enable it for these tests.

The rest of the tests are disabled on a per-test basis.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>