charm_builder: build dependencies locally
Comply with the requirement of not installing binary dependencies,
building them locally instead. The process may take a long time to
complete, so use the venv module to install requirements to a persistent
staging virtualenv so that dependencies already installed won't be
reinstalled (and rebuilt) when the charm is repacked.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>