charm_builder: install binary packages with no special options
Option `--prefer-binary` is only available in pip 20.x and bionic
installs 9.x. Options are to update pip to a newer version and keep
using `--prefer-binary`; change the option to `--only-binary` (in
this case installation of packages only available in source form
or depending on packages only available in source form would fail); or
not specify any option related to binary installation (which would
result in installation from source if tha latest version of a package
is only available in source form, which is a possible situation but
likely to be a corner case). This commit implements the third
solution.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>