store: handle store errors in whoami command
Define a `allow_login` parameter to the store client decorator so
that the `whoami` command can use it to handle errors without
automatically logging the user in.

Example error messages:

- If no external credentials are used, and the user is not logged in:
```
$ charmcraft whoami
You are not logged in to Charmhub.
```

- If credentials are used, and the response code is 401:
```
$ CHARMCRAFT_AUTH="$(cat ~/creds-expired)" charmcraft whoami
Provided credentials are no longer valid for Charmhub. Regenerate them and try again.
```

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>