Add name argument for upload command
In some cases the name of the charm a user wants to upload is not the
same as the one specified in metadata.yaml. Specifically for charms
imported into charmhub from the old charmstore this isn't true if they
formerly namespaces (e.g. cs:~foo/bar become foo-bar but charm name
remains foo in metadata.yaml).

To allow charms to be still uploaded in this case a --name option is
added to the upload command which allows a user to specify the exact
name of the charm he wants to upload to rather than letting charmcraft
automatically extract the name from the metadata.yaml file

Fixes #681