Manifest Linter
The following implements a manifest linter that ensures that you don't
have both a metadata series and a manifest base. By doing this, we
can make the Juju side of the code simpler in the future. If all with in
the charmhub store and locally deployed charms with Juju only have one
source of truth for the series information it makes determination easy.

Additionally, Juju wants to move away from series terminology and this
will help with that, as in the future we would want to deprecate
metadata.series field.