Enforced DN03: the 'bases' is now mandatory for charms in the config.
Of course this not only led to the change in the code, but to a lot of
config not-bases-related tests where I had to add 'bases'. Note that in
a lot of cases I just changed the 'type' to 'bundle' (where being a
charm was not relevant to the test) to minimize the amount of changes.