Finally isolated charm_builder.py as a generic plugin script.
This will stop its execution from showing some weirdness in the logs and improve the general structure of the code (stop mixing code run from the plugin with the rest of Charmcraft).

After a long talk with Claudio yesterday we decided to convert all outputs to plain prints to stdout, again, as a regular plugin script.

Also converted the CraftError exceptions to Runtime ones (fully isolation from Craft CLI) and improved the message.