Support both image id and digest when uploading oci-image resources.
Also be very explicit on which one is used when. It results on being quicker if the id is used (no check on remote registry) and better messages to the user. Which were improved also: not all needed to be permanent, and actually ended in error if the image to upload is not found and nothing can be done.

Fixes #853.