Deprecate 'series' in metadata.yaml & patch linter
Patch the metadata linter to fail if the 'series' key is present. For a
long time charmcraft has generated a manifest.yaml on pack which has
been used instead series in metadata.

Juju ignores 'series' in metadata for any packed charm (i.e. a charm with
a manifest.yaml)

There is an effort to remove the notion of a series from Juju entirely,
substituting it for bases. As such, Juju is roadmapped to deprecate support
for charms without manifest.yaml files in 3.4, removing support entirely in
4.0 (i.e. showing an error).

Showing a linter error in `charmcraft analyze` will be a convenient way
to begin communicating this

As a flyby fix, prune whitespace in template files