fix(charm_builder): bump minimum pip version to 23 (#1437)
fixes #1374

Pip 22 doesn't use the prebuilt wheel cache when using
`--no-binary=[list of packages]`, but does when using
`--no-binary=:all:`. As such, using strict dependencies won't use the
cache with pip 22

Has potential regressions, so I'm going to run a large spread on this:
https://github.com/canonical/charmcraft/actions/runs/7184537100

EDIT: Large spread failures are unrelated and are fixed in
https://github.com/canonical/charmcraft/pull/1438. This should be
cherry-picked to the 2.5 branch.