Add the `flask-framework` extension (#1582)
Bring the `flask-framework` extension from the `feature/12f` branch into
the main branch.

The `flask-extension` is a part of the 12-factor toolchain that assists
users in creating a charm to manage their Flask application within the
Juju ecosystem.

As we are still awaiting some changes in other libraries to be
finalized, please review the current content. I will update the pull
request once the changes in the other libraries have been merged.