feat(config): support `type:secret` in config.options (#1625)
Since Juju 3.1, the `config.options` section of `charmcraft.yaml` (or
the older `config.yaml`) has supported a `type` value of `secret` (for
user secrets). This is missing from the documentation right now, but the
Juju team are working on fixing that.

This PR adds support for `type: secret` in config options.

---------

Co-authored-by: Alex Lowe <alex.lowe@canonical.com>