feat: flask and django, update to paas-app-charmer 1.0.3 (#1691)
Version 1.0.3 of
[paas-app-charmer](https://github.com/canonical/paas-app-charmer) has
been released, providing the new following functionality:

- Two new integrations, SAML and S3.
- Blocking on non optional integrations. If the optional attribute in
the `charmcraft.yaml` file is set to false and the integration is not
active, paas-app-charmer will block the charm (and stop all the
services).

This PR adapts charmcraft to get advantage of that functionality.

Besides those changes, `charm-libs` has been added to the extension
(including the new two new libraries for SAML and S3). That way the user
can do `charmcraft fetch-libs` instead of one `fetch-lib` per library.