feat: go-framework init profile and extension (#1800)
This PR implements the spec ISD153, "12-Factor Go Support" for
charmcraft.

For that, a new init profile called "go-framework" and a new go
extension, called "go-framework" are needed. These additions are similar
to the flask and django init profiles and extensions.

For the Go profile, the  base 24.04 is used.

Code in extensions is refactored, so all 12 factor extensions, that are
similar, inherit from _TwelveFactorBase.