feat: use ops.main() type hints (#1911)
The `ops.main` entry point will be properly typed starting with the
upcoming release, ops==2.17.0

This PR updates the templates so that new charms use the fixed type
hints.

It is paired with #1883 where charmcraft analyse validates that ops.main
is called.

We'll be relying on charm repo static type checker to validate that
ops.main is called with correct arguments.

- [x] wait for ops 2.17.0 to be released
- [x] update ops version in the templates
- [x] update ops.main() call in the templates

https://warthogs.atlassian.net/browse/CHARMTECH-276
parent: https://warthogs.atlassian.net/browse/CHARMTECH-219

---------

Co-authored-by: Alex Lowe <alex.lowe@canonical.com>