Rename paas-app-charmer -> paas-charm (#1961)
The library
[`paas-app-charmer`](https://github.com/canonical/paas-app-charmer) has
been renamed to [`paas-charm`](https://github.com/canonical/paas-charm).

The change is backward compatible, as last version of the
`paas-app-charmer` will be empty and will point to the new version (and
will keep the old entrypoint classes).

This PR changes the references from paas-app-charm to paas-charm and
also the main charm module to the new location (`paas_app_charmer`
module renamed to `paas_charm`).