fix(snap): check if LXD is actually installed
On recent Ubuntu systems, "lxc" might be "/usr/sbin/lxc", which is provided
by the "lxd-installer" package and will install the LXD snap if it's not
installed. This installation can then take a long time if the store is having
issues.

For the purposes of the configure and remove hooks we *don't* want to install
LXD just to check that it has no stale images, so update the hooks to do some
early detection and bail out if LXD is not installed.

Fixes #1982