build(deps): update craft-application and craft-providers (#2069)
Use a unique build environment for each platform in a charm recipe.

### Scenario 1

* Fixes a bug where multiple platforms would build in the same build
environment if the platforms had the same `build-on` and `build-for`
entries.

```yaml
platforms:
  noble:
      build-on: [riscv64]
      build-for: [riscv64]
  noble-debug:
      build-on: [riscv64]
      build-for: [riscv64]
```

###  Scenario 2

* Improves build time for multi-base recipes. Previously, Charmcraft
would tear down and recreate the same container when building for a
different base if the platforms had the same `build-on` and `build-for`
entries.

```yaml
platforms:
  jammy:
      build-on: [ubuntu@22.04:riscv64]
      build-for: [ubuntu@22.04:riscv64]
  noble:
      build-on: [ubuntu@24.04:riscv64]
      build-for: [ubuntu@24.04:riscv64]
```


Fixes #2058 
(CRAFT-3847)

Signed-off-by: Callahan Kovacs <callahan.kovacs@canonical.com>