docs: add, refactor, and reformat discourse docs feat. charmcraft (#2010)
This PR adds all the juju.is/docs content featuring Charmcraft.

The content was originally formatted for Discourse; this PR reformats it
to serve the RTD project the docs are destined for.

The content was also originally entangled with Ops content; this PR
significantly refactors -- as well as clarifies and tries to complete --
everything to serve the stand-alone Charmcraft docs set it is destined
for.

Issues that should be fixed in subsequent PRs:

- Everywhere: (1) (a) Links to juju.is will work, as even after moving
docs out of there we'll have redirects; however, they should be updated.
(b) Links to ops docs are empty -- they reference content I'm planning
to move there post-refactoring; once we have them in place, which should
be in the next few days and pre-official-Charmcraft-RTD-launch, we
should add them in Charmcraft docs.
- Tutorial: The tutorial should also be offered in the "regular" flavor,
that is, for apps that are not 12-factor. That flavor should also
feature the publish step. (See, e.g., our old Kubernetes tutorial, which
includes https://juju.is/docs/sdk/publish-your-charm-on-charmhub . PS If
the Charmcraft tutorial starts featuring this, as it should, the old
Kubernetes tutorial, which will be moved to the stand-alone Ops docs,
can focus on just development, which is what Ops is really all about.)
- How-to guides: While this PR surfaces a lot more of Charmcraft and
arguably a lot more clearly, there are still corners with cobwebs or
gaps. These include: (1) Manage charms (12-factor app) -- I think this
doc should be incorporated into the main Manage charms doc, which could
have tabs for "regular', "django", etc. (This would align with plans for
the juju.is/docs entrypoint into the "federated docs"). (2) The Manage
parts doc, which is full of "TBA" -- i.e., has gaps. (3) The docs in
"Misc", which are not well integrated with everything else here. (4)
Maybe the Manage bundles doc, which doesn't by far tell the full story;
given that we're phasing bundles out, it's also perhaps fine not to
worry about this too much. (5) Commands `analyse` and `test` are not
featured in any how-to guide.
- Reference: (1) The autogenerated `charmcraft` CLI command reference
docs should have anchors on the template `command-charmcraft-x` and
titles on the template "Command `charmcraft x`. This will not better
align with existing practices in the Juju world but also results in more
clarity when linking to the page using `{ref}`...`` (2) The various file
reference docs should be as much as possible autogenerated from source
(likely the source will have to be updated first) -- at the very least,
file `charmcraft.yaml`. They may also need further clean-up. (3) The
part docs are a bit of a mess. We should move any relevant content to
the part docs, then link up to that and only keep in Charmcraft the
material specific to Charmcraft docs, and then revisit that material as
well to streamline it as much as possible (e.g., currently the dump
plugin is documented in 3 places). (4) Charmcraft analyzers and linters
-- this gives details on pack and analyse but it's not clear in what way
it's useful to the reader, so we need to rethink.
- Explanation: We have nothing there. I didn't delete that section as I
think there is potentially a topic we could have there -- e.g.,
Charmcraft in the bigger context of Starcraft (a clarification of the
general design philosophy there and trends, maybe).

PS Some docs may have a Contributors list on the bottom, listing the
Discourse handles of the contributors to date. We'll want to rethink how
we do that.

---------

Co-authored-by: Erin Conley <erin.conley@canonical.com>
Co-authored-by: Michael DuBelko <michael.dubelko@gmail.com>
Co-authored-by: Alex Lowe <alex.lowe@canonical.com>
Co-authored-by: Michael DuBelko <michael.dubelko@canonical.com>