feat: auto-generate charmcraft.yaml json schema (#2183)
Adds plumbing for creating a JSON schema for project files, a CI
workflow to check for changes, and a Makefile recipe to regenerate the
schema.

Once this is merged, I will have to have a PR in the schema store to
point to this repository be merged. Until then, the only way to test
this is to locally point your editor to `schema/charmcraft.json` for
`**/charmcraft.yaml`.

Corresponding schema store PR:
https://github.com/SchemaStore/schemastore/pull/4472

CRAFT-2806.