feat(extension): Block on missing required config (#2187)
This PR is related to 12 Factor [Block on missing required config
PR](https://github.com/canonical/paas-charm/pull/42)

## Rationale

Some applications require certain config options to run and would crash
if these options are not provided. To prevent this, we need to add an
optional field to the config options and block the charm if a config
option marked as non-optional has not been provided with a message that
explains which config options are missing.

To stay backward compatible the default value for the optional field
will be true.

Since it doesn’t make sense for config options that are non-optional to
have a default value, charmcraft will look for this condition at pack
time and error out.


## Extension Changes

Adds a check for non-optional options with default values and throws
`ExtensionError` when found.
Adds documentation about the non-optional options.


Reviewers:
@jdkandersson 
@javierdelapuente 
@erinecon