docs: add `charm-user` in the charmcraft.yaml file reference (#2195)
The `charm-user` field is documented in the metadata.yaml file
reference, but is missing from the charmcraft.yaml file reference, even
though it does work there. This PR adds it to that reference.

As the docs currently are, it gives the incorrect impression that you
cannot use the combined charmcraft.yaml file to specify `charm-user`,
and must have separate metadata, actions, and config files.

---------

Co-authored-by: Erin Conley <erin.conley@canonical.com>
Co-authored-by: Ali UĞUR <39213991+alithethird@users.noreply.github.com>