Docs: Update 12-factor tutorials (#2241)
Updates to the 12-factor tutorials that affect all four tutorials.

### Summary of changes
* Add a tip admonition to let the user know about `charmcraft
expand-extensions`. This has been a repeated ask from UX research
participants.
* Convert the `juju status` admonition to a tip to recommend that users
start a `juju status` session in a new tab to monitor their deployment.
This prepares the user for a successful deployment setup.
* Update the terminal output for `rockcraft pack` to use a generic
architecture stand-in. This indicates that the architecture plays an
active role in defining the rock.
* Mention that subsequent `charmcraft pack` and `rockcraft pack`
commands are faster after the initial packing. This has been a repeated
ask from UX research participants.
* Condense text about adding an architecture constraint to the Juju
model, for conciseness.