Docs: 12-factor tutorial updates (#2257)
Six updates proposed by @evildmp in
https://github.com/canonical/charmcraft/issues/2235 that affect all four
of the 12-factor tutorials.

### Summary of changes
* Converted first note admonition into a normal paragraph.
* Updated the ambiguity of "Ensure that platforms includes the
architecture of your host." by using the suggestion "Platforms must
match the architecture of your host."
* Updated the instruction to edit `rockcraft.yaml` using the suggestion.
* Removed an explicit reference to AMD64 architecture at the Juju model
constraint step to let the user build up muscle memory. Architecture is
still mentioned because that's what the command is explicitly doing.
* (Django tutorial only) Updated "The generated Django app doesn’t come
with an app" to reference the Django *project* instead.
* (Django tutorial only) Updated `MULTIPASS_PRIVATE_IP` to use regular
text formatting, as the original spelling implied a variable.