docs: 12-Factor tutorial updates (#2263)
Five updates proposed by @evildmp in
https://github.com/canonical/charmcraft/issues/2235 that affect all four
of the 12-factor tutorials and the reusable setup scripts.

### Summary of changes
* Updated the Multipass install link to point to the correct how-to
guide.
* Removed explicit mentions of "If you aren't on a host with AMD64
architecture" and the note admonitions around the rock and charm names.
This is to help the user build up the muscle memory of the naming and
take responsibility for the file names.
* Added a brief description of the ``--profile`` flag in the rock and
charm creation sections.
* (Django tutorial only) Reduced the text in the "We'll need to update
the settings.py file..." paragraph, again to help the user build muscle
memory of where the file is located.
* (Django tutorial only) Rephrased the text about connecting the Django
app to PostgreSQL, as that's not what's actually going on at that time.
Instead we're adding the PostgreSQL relation endpoint to the charm
project file, so the text now reflects that change.

In addition to these changes, a warning admonition block about older
versions of Charmcraft was removed in the Django tutorial.

---------

Signed-off-by: Alex Lowe <alex.lowe@canonical.com>
Co-authored-by: Alex Lowe <alex.lowe@canonical.com>