docs: Updates to the 12-factor tutorials (#2276)
Six updates proposed by @evildmp in
https://github.com/canonical/charmcraft/issues/2235 that affect all four
of the 12-factor tutorials.

### Summary of changes
* Added a brief explanation of a rock at the beginning of the rock
sections.
* (Django tutorial) Before packing the rock, added a note that explains
that the user can no longer run or test the app locally due to the
changes we made in preparing the app for a PostgreSQL database.
* (FastAPI and Go tutorials) Changed some instances of the mechanical
instruction "open" to a functional action "edit" for updating the
`charmcraft.yaml` file. This also establishes consistency with the
corresponding rock sections.
* Added a `seealso` box to `charmcraft pack` to provide more context to
the user.
* Added a brief explanation of a Juju model at the beginning of the
deploy section.
* (Django tutorial) When setting the `django-debug` configuration, added
more information in the note admonition that this configuration sets an
environment variable in the Django app. This gives the user a way to
connect the Juju-specific command with a familiar concept.