docs: copy best practices from Juju docs (#2203)
This PR migrates best practices from [the Juju
docs](https://canonical-juju.readthedocs-hosted.com/en/latest/user/reference/charm/charm-development-best-practices/),
where the best practice applies to Charmcraft. It is paired with a
[similar PR for ops](https://github.com/canonical/operator/pull/1615).

The intention is to standardise best practices so that the docs can be
scraped to collect best practices across Ops, Charmcraft, juju, and
elsewhere in the charming ecosystem, to provide a comprehensive
checklist for charmers.

My suggestion is for ReST we use:

```
.. admonition:: Best practice
    :class: hint

    Content goes here
```

This provides an admonition box that has the title "Best Practice", but
is otherwise styled as a hint (David Wilding is working on putting a
custom icon into the starter pack, to replace the ? icon that Furo
defaults to).

The best practices here were selected in a review by @tmihoc and
@dwilding and myself, although all problems with the suggested wording
are my fault alone.

---------

Co-authored-by: Erin Conley <erin.conley@canonical.com>
Co-authored-by: Ali UĞUR <39213991+alithethird@users.noreply.github.com>
Co-authored-by: Alex Lowe <alex.lowe@canonical.com>
Co-authored-by: Michael DuBelko <michael.dubelko@gmail.com>
Co-authored-by: Teodora Mihoc <teodora.mihoc@canonical.com>