docs: Updates to the 12-factor tutorials (#2282)
Three updates proposed by @evildmp (two updates in
https://github.com/canonical/charmcraft/issues/2235 plus one other) that
affect all four tutorials and their `task.yaml` files.

### Summary of changes
* Added an admonition for `rockcraft pack` to provide more context to
the user.
* Provided more information for the `rockcraft.skopeo copy` command to
give a better sense of what each of the options and tags are doing.
Updated the `task.yaml` files so that the command appears over more
lines, and changed the `seealso` admonition to a `note` block to provide
a brief description of each part while also including the external link.
* Updated the description of the MicroK8s registry to describe the
registry as a temporary Dockerhub.

---------

Co-authored-by: Alex Lowe <alex.lowe@canonical.com>