docs: update 12-factor how-to guide on async apps (#2352)
Relevant ticket: ISD-3593

Maintenance work on the how-to guide "Write a Kubernetes charm for an
async Flask app".

### Summary of work
* Add a note admonition that the guide also applies to Django apps.
* Add a caption to the `requirements.txt` file included in the guide.
* Add in the assumption that the app name is `flask-async-app`.
* Originally, we had the suggestion to add `juju status` output to the
guide. Since the configuration is not represented in the `juju status`
table, I decided it wasn't worth adding in. However, I did add a
suggestion to monitor the app status while waiting for the configuration
to take effect.

Co-authored-by: Michael DuBelko <michael.dubelko@canonical.com>