docs: fix broken link to Juju docs in template CONTRIBUTING.md (#2350)
juju.is/docs/sdk/dev-setup 404s now. This PR changes the links to that
page in the template CONTRIBUTING.md to point to [the current
equivalent](https://github.com/tonyandrewmeyer/charmcraft/pull/new/fix-link-in-contributing.md).

(There are many other juju.is links, but fixing these ones in particular
seems straightforward, and many of the other redirect).