fix: correct the container name in the 'simple' profile (#2394)
The 'simple' project was updated to use a different container name in the charm code and tests but the charmcraft.yaml name was not updated.

When run in 3.5 a freshly init'd "simple" will fail tests, with this fix they pass.

The "simple" profile has been removed in main, but it'd be nice to have it work correctly in 3.5.