docs: add related links section to docs homepage (#2464)
This PR adds `relatedlinks` metadata to the Charmcraft documentation
site homepage.

This uses the related links extension included in Canonical's Sphinx
extensions.

I've used the same list that we use for
[ops](https://documentation.ubuntu.com/ops/latest/) and
[charmlibs](https://documentation.ubuntu.com/charmlibs/), with the
addition of Craft Parts. Perhaps there are more craft specific links
that could be included, or maybe some of the currently included links
aren't related enough (Jubilant? Pebble?). The list is currently
(manually) sorted alphabetically, but this can be changed too.

---------

Signed-off-by: James Garner <james.garner@canonical.com>
Co-authored-by: Michael DuBelko <michael.dubelko@canonical.com>